using vagrant
Here the host is an Ubuntu box. In this asciicast I demonstrate how to use Vagrant. We download an Ubuntu 12.04 image and log into it. In the VM we create a file in the shared folder /vagrant. When we log out of the VM, the created file is there in the shared folder. Finally, the running VM is stopped.